Public

+++ OracleShellInstall.sh
@@ -4,7 +4,7 @@ echo "##Author 	: LuciferLiu"
 echo "##Blog   	: https://blog.csdn.net/m0_50546016"
 echo "##Github        : https://github.com/pc-study/InstallOracleshell"
 echo "##Version	: 1.0"
-echo "##Function   	: Oracle 11g/12c/18c/19c(Single and Rac) install on Linux 6/7"
+echo "##Function   	: Oracle 11g/12c/18c/19c(Single and Rac) install on Linux 6/7/8"
 echo "####################################################################################"
 echo "#执行脚本前："
 echo "#    1. 把脚本放入软件目录，例如：/soft"
@@ -60,7 +60,7 @@ OCRREDUN=EXTERNAL
 DATAREDUN=EXTERNAL
 num1=0
 num2=0
-TIMESERVER=
+TIMESERVERIP=
 ONLYCONFIGOS=N
 ONLYINSTALLGRID=N
 ONLYINSTALLORACLE=N
@@ -279,7 +279,7 @@ help() {
   c1 "-od,		--OCRP_BASEDISK			RAC OCRDISK DISKNAME" green
   c1 "-or,		--OCRREDUN			RAC OCR REDUNDANCY(EXTERNAL|NORMAL|HIGH)" green
   c1 "-dr,		--DATAREDUN			RAC DATA REDUNDANCY(EXTERNAL|NORMAL|HIGH)" green
-  c1 "-ts,            --TIMESERVER                    RAC TIME SERVER IP" green
+  c1 "-tsi,            --TIMESERVERIP                    RAC TIME SERVER IP" green
   c1 "-txh            --TuXingHua                     Tu Xing Hua Install" green
   c1 "-udev           --UDEV                          Whether Auto Set UDEV" green
   c1 "-dns            --DNS                           RAC CONFIGURE DNS(Y|N)" green
@@ -442,8 +442,8 @@ while [ -n "$1" ]; do #Here by judging whether $1 exists
     RACPRIVFCNAME1=$2
     shift 2
     ;;
-  -ts | --TIMESERVER)
-    TIMESERVER=$2
+  -tsi | --TIMESERVERIP)
+    TIMESERVERIP=$2
     shift 2
     ;;
   -node | --nodeNum)
@@ -614,8 +614,8 @@ if [ "${nodeNum}" -eq 1 ]; then
         echo -e " -dns ${DNS}\c"
       } >"${SOFTWAREDIR}"/racnode2.sh
       ##TimeServer
-      if [ -n "${TIMESERVER}" ]; then
-        echo -e " -ts ${TIMESERVER}\c" >>"${SOFTWAREDIR}"/racnode2.sh
+      if [ -n "${TIMESERVERIP}" ]; then
+        echo -e " -ts ${TIMESERVERIP}\c" >>"${SOFTWAREDIR}"/racnode2.sh
       fi
 
       ##Two Private ip
@@ -660,6 +660,8 @@ if [ "$OS_VER_PRI" -eq 7 ]; then
   OS_VERSION=linux7
 elif [ "$OS_VER_PRI" -eq 6 ]; then
   OS_VERSION=linux6
+elif [ "$OS_VER_PRI" -eq 8 ]; then
+  OS_VERSION=linux8
 else
   c1 "sorry, this operating system is not supported!!!" red
   exit 99
@@ -814,13 +816,28 @@ InstallRPM() {
     exit 99
   else
     if [ ! -f /etc/yum.repos.d/local.repo ]; then
-      {
-        echo "[server]"
-        echo "name=server"
-        echo "baseurl=file://""${mountPatch}"
-        echo "enabled=1"
-        echo "gpgcheck=1"
-      } >/etc/yum.repos.d/local.repo
+      if [ "${OS_VERSION}" = "linux6" ] || [ "${OS_VERSION}" = "linux7" ]; then
+        {
+          echo "[server]"
+          echo "name=server"
+          echo "baseurl=file://""${mountPatch}"
+          echo "enabled=1"
+          echo "gpgcheck=1"
+        } >/etc/yum.repos.d/local.repo
+      elif [ "${OS_VERSION}" = "linux8" ]; then
+        {
+          echo "[BaseOS]"
+          echo "name=BaseOS"
+          echo "baseurl=file:///${mountPatch}/BaseOS"
+          echo "enabled=1"
+          echo "gpgcheck=1"
+          echo "[AppStream]"
+          echo "name=AppStream"
+          echo "baseurl=file:///${mountPatch}/AppStream"
+          echo "enabled=1"
+          echo "gpgcheck=1"
+        } >/etc/yum.repos.d/local.repo
+      fi
       rpm --import "${mountPatch}"/RPM-GPG-KEY-redhat-release
     fi
     if [ "${OS_VERSION}" = "linux6" ]; then
@@ -865,9 +882,9 @@ InstallRPM() {
           ksh \
           nfs-utils --skip-broken
       fi
-    elif [ "${OS_VERSION}" = "linux7" ]; then
+    elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
       if [ "${TuXingHua}" = "y" ] || [ "${TuXingHua}" = "Y" ]; then
-        #LINUX 7
+        #LINUX 7 && LINUX 8
         yum groupinstall -y "Server with GUI"
         yum install -y tigervnc*
       fi
@@ -908,6 +925,13 @@ InstallRPM() {
           openssh-clients \
           readline* \
           psmisc --skip-broken
+      fi
+      ##Solutions: error while loading shared libraries: libnsl.so.1: cannot open shared object
+      ##Requirements for Installing Oracle Database/Client 19c on OL8 or RHEL8 64-bit (x86-64) (Doc ID 2668780.1)
+      if [ "${OS_VERSION}" = "linux8" ]; then
+        dnf install -y librdmacm
+        dnf install -y libnsl*
+        dnf install -y libibverbs
       fi
     fi
 
@@ -957,12 +981,14 @@ EOF
           fi
         fi
       fi
-      if rpm -ivh "${SOFTWAREDIR}"/compat-libstdc++-33-3.2.3-72.el7.x86_64.rpm; then
-        if [ "${OracleInstallMode}" = "rac" ] || [ "${OracleInstallMode}" = "RAC" ]; then
-          scp "${SOFTWAREDIR}"/compat-libstdc++-33-3.2.3-72.el7.x86_64.rpm "${RAC2PUBLICIP}":/root
-          ssh "$RAC2PUBLICIP" rpm -ivh /root/compat-libstdc++-33-3.2.3-72.el7.x86_64.rpm
+      if [ -f "${SOFTWAREDIR}"/compat-libstdc++-33-3.2.3-72.el7.x86_64.rpm ]; then
+        if rpm -ivh "${SOFTWAREDIR}"/compat-libstdc++-33-3.2.3-72.el7.x86_64.rpm; then
+          if [ "${OracleInstallMode}" = "rac" ] || [ "${OracleInstallMode}" = "RAC" ]; then
+            scp "${SOFTWAREDIR}"/compat-libstdc++-33-3.2.3-72.el7.x86_64.rpm "${RAC2PUBLICIP}":/root
+            ssh "$RAC2PUBLICIP" rpm -ivh /root/compat-libstdc++-33-3.2.3-72.el7.x86_64.rpm
+          fi
+          rm -rf "${SOFTWAREDIR}"/compat-libstdc++-33-3.2.3-72.el7.x86_64.rpm
         fi
-        rm -rf "${SOFTWAREDIR}"/compat-libstdc++-33-3.2.3-72.el7.x86_64.rpm
       fi
     fi
   fi
@@ -982,7 +1008,7 @@ EOF
 
   if [ "${OS_VERSION}" = "linux6" ]; then
     logwrite "RPM Check" "rpm -q bc binutils compat-libcap1 compat-libstdc++-33 gcc gcc-c++ elfutils-libelf elfutils-libelf-devel glibc glibc-devel libaio libaio-devel libgcc libstdc++ libstdc++-devel libxcb libX11 libXau libXi libXrender make net-tools smartmontools sysstat e2fsprogs e2fsprogs-libs expect unzip openssh-clients readline"
-  elif [ "${OS_VERSION}" = "linux7" ]; then
+  elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
     logwrite "RPM Check" "rpm -q bc binutils compat-libcap1 compat-libstdc++-33 gcc gcc-c++ elfutils-libelf elfutils-libelf-devel glibc glibc-devel ksh libaio libaio-devel libgcc libstdc++ libstdc++-devel libxcb libX11 libXau libXi libXtst libXrender libXrender-devel make net-tools nfs-utils smartmontools sysstat e2fsprogs e2fsprogs-libs fontconfig-devel expect unzip openssh-clients readline"
   fi
 }
@@ -997,7 +1023,7 @@ SetHostName() {
       /bin/hostname "$hostname"
       sed -i "s/${Hostname}/HOSTNAME=${hostname}/" /etc/sysconfig/network
     fi
-  elif [ "${OS_VERSION}" = "linux7" ]; then
+  elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
     if [[ $(grep -E "${hostname}" /etc/hostname) != "${hostname}" ]]; then
       /usr/bin/hostnamectl set-hostname "${hostname}"
     fi
@@ -1096,7 +1122,7 @@ DNSServerConf() {
   yum install -y bind-libs bind bind-utils
   if [ "${OS_VERSION}" = "linux6" ]; then
     chkconfig named on
-  elif [ "${OS_VERSION}" = "linux7" ]; then
+  elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
     systemctl enable named
   fi
   if [ "$nodeNum" -eq 1 ]; then
@@ -1444,7 +1470,7 @@ CreateUsersAndDirs() {
   ## Judge DISK SPACE
   if [ "${OS_VERSION}" = "linux6" ]; then
     BASEDIR_SPACE=$(df "${ENV_BASE_DIR}" | tail -n 1 | awk '{printf $3}')
-  elif [ "${OS_VERSION}" = "linux7" ]; then
+  elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
     BASEDIR_SPACE=$(df "${ENV_BASE_DIR}" | tail -n 1 | awk '{printf $4}')
   fi
   BASEDIR_SPACE=$((BASEDIR_SPACE / 1024 / 1024))
@@ -1530,7 +1556,7 @@ EOF
   alias ocr_${num1}
   }
 EOF
-    elif [ "${OS_VERSION}" = "linux7" ]; then
+    elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
       cat <<EOF >>/etc/multipath.conf
   multipath {
   wwid "$(/usr/lib/udev/scsi_id -g -u "${i}")"
@@ -1579,7 +1605,7 @@ EOF
   alias data_${num2}
   }
 EOF
-    elif [ "${OS_VERSION}" = "linux7" ]; then
+    elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
       cat <<EOF >>/etc/multipath.conf
   multipath {
   wwid "$(/usr/lib/udev/scsi_id -g -u "${i}")"
@@ -1622,7 +1648,7 @@ EOF
 
   if [ "${OS_VERSION}" = "linux6" ]; then
     start_udev
-  elif [ "${OS_VERSION}" = "linux7" ]; then
+  elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
     udevadm control --reload-rules
     udevadm trigger --type=devices
   fi
@@ -1711,7 +1737,7 @@ EOF
     logwrite "ntpd" "service ntpd status"
 
     [ -f /etc/ntp.conf ] && mv /etc/ntp.conf /etc/ntp.conf.orig
-  elif [ "${OS_VERSION}" = "linux7" ]; then
+  elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
     timedatectl set-timezone Asia/Shanghai
     if [ "$(systemctl status chronyd | grep -c running)" -gt 0 ]; then
       systemctl stop chronyd.service
@@ -1720,22 +1746,36 @@ EOF
 
     logwrite "chronyd" "systemctl status chronyd"
   fi
-
   ##ntpdate configure
-  if [[ -n $TIMESERVER ]]; then
-    yum install -y ntpdate
-    if [ ! -f /var/spool/cron/root ]; then
-      echo "##For ntpupdate" >>/var/spool/cron/root
+  if [[ -n "${TIMESERVERIP}" ]]; then
+    if [ "${OS_VERSION}" = "linux6" ] || [ "${OS_VERSION}" = "linux7" ]; then
+      yum install -y ntpdate
+      if [ ! -f /var/spool/cron/root ]; then
+        echo "##For ntpupdate" >>/var/spool/cron/root
+      fi
+      if [ "$(grep -E -c "#OracleBegin" /var/spool/cron/root)" -eq 0 ]; then
+        [ ! -f /var/spool/cron/root."${DAYTIME}" ] && cp /var/spool/cron/root /var/spool/cron/root."${DAYTIME}" >/dev/null 2>&1
+        {
+          echo "#OracleBegin"
+          echo "00 12 * * * /usr/sbin/ntpdate -u ${TIMESERVERIP} && /usr/sbin/hwclock -w"
+          echo "#OracleEnd"
+        } >>/var/spool/cron/root
+      fi
+      /usr/sbin/ntpdate -u "${TIMESERVERIP}" && /usr/sbin/hwclock -w
+    elif [ "${OS_VERSION}" = "linux8" ]; then
+      if [ ! -f /var/spool/cron/root ]; then
+        echo "##For ntpupdate" >>/var/spool/cron/root
+      fi
+      if [ "$(grep -E -c "#OracleBegin" /var/spool/cron/root)" -eq 0 ]; then
+        [ ! -f /var/spool/cron/root."${DAYTIME}" ] && cp /var/spool/cron/root /var/spool/cron/root."${DAYTIME}" >/dev/null 2>&1
+        {
+          echo "#OracleBegin"
+          echo "00 12 * * * /usr/sbin/chronyd -q "server "${TIMESERVERIP}" iburst" && timedatectl set-local-rtc true"
+          echo " #OracleEnd"
+        } >>/var/spool/cron/root
+      fi
+      chronyd -q "server ${TIMESERVERIP} iburst" && timedatectl set-local-rtc true
     fi
-    if [ "$(grep -E -c "#OracleBegin" /var/spool/cron/root)" -eq 0 ]; then
-      [ ! -f /var/spool/cron/root."${DAYTIME}" ] && cp /var/spool/cron/root /var/spool/cron/root."${DAYTIME}" >/dev/null 2>&1
-      {
-        echo "#OracleBegin"
-        echo "00 12 * * * /usr/sbin/ntpdate -u ${TIMESERVER} && /usr/sbin/hwclock -w"
-        echo "#OracleEnd"
-      } >>/var/spool/cron/root
-    fi
-    /usr/sbin/ntpdate -u "${TIMESERVER}" && /usr/sbin/hwclock -w
     logwrite "Time ntpdate" "crontab -l"
   fi
 
@@ -1755,15 +1795,15 @@ Disableavahi() {
     fi
     logwrite "avahi-daemon" "service avahi-daemon  status"
 
-  elif [ "${OS_VERSION}" = "linux7" ]; then
+  elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
     yum install -y avahi*
     if [ "$(systemctl status avahi-daemon | grep -c running)" -gt 0 ]; then
       systemctl stop avahi-daemon.socket
-      systemctl disable avahi-daemon.socket
       systemctl stop avahi-daemon.service
-      systemctl disable avahi-daemon.service
       pgrep -f avahi-daemon | awk '{print "kill -9 "$2}'
     fi
+    systemctl disable avahi-daemon.service
+    systemctl disable avahi-daemon.socket
     logwrite "avahi-daemon" "systemctl status avahi-daemon"
   fi
 }
@@ -1780,7 +1820,7 @@ DisableFirewall() {
       chkconfig ip6tables off
     fi
     logwrite "Iptables" "service iptables status"
-  elif [ "${OS_VERSION}" = "linux7" ]; then
+  elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
     if [ "$(systemctl status firewalld.service | grep -c running)" -gt 0 ]; then
       systemctl stop firewalld.service
       systemctl disable firewalld.service
@@ -1824,7 +1864,7 @@ EOF
       sed -i 's/quiet/quiet numa=off/' /boot/grub/grub.conf
     fi
     logwrite "/boot/grub/grub.conf" "cat /boot/grub/grub.conf"
-  elif [ "${OS_VERSION}" = "linux7" ]; then
+  elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
     if [ "$(grep -E -c "transparent_hugepage=never numa=off" /etc/default/grub)" -eq 0 ]; then
       [ ! -f /etc/default/grub."${DAYTIME}" ] && cp /etc/default/grub /etc/default/grub."${DAYTIME}"
       sed -i 's/quiet/quiet transparent_hugepage=never numa=off/' /etc/default/grub
@@ -1849,7 +1889,7 @@ DisableNetworkManager() {
       service NetworkManager off
     fi
     logwrite "NetworkManager" "service NetworkManager status"
-  elif [ "${OS_VERSION}" = "linux7" ]; then
+  elif [ "${OS_VERSION}" = "linux7" ] || [ "${OS_VERSION}" = "linux8" ]; then
     #Turn off the NetworkManager(Linux 7)
     if [ "$(systemctl status NetworkManager.service | grep -c running)" -gt 0 ]; then
       systemctl stop NetworkManager.service
@@ -1892,9 +1932,19 @@ net.core.rmem_default = 262144
 net.core.rmem_max = 4194304
 net.core.wmem_default = 262144
 net.core.wmem_max = 1048576
+EOF
+    if [ "${OS_VERSION}" = "linux8" ]; then
+      cat <<EOF >>/etc/sysctl.conf
+# sysctl kernel.numa_balancing
+kernel.numa_balancing = 0
+EOF
+    fi
+    if [ -n "${RACPUBLICFCNAME}" ] && [ -n "${RACPRIVFCNAME}" ]; then
+      cat <<EOF >>/etc/sysctl.conf
 net.ipv4.conf.${RACPUBLICFCNAME}.rp_filter = 1
 net.ipv4.conf.${RACPRIVFCNAME}.rp_filter = 2
 EOF
+    fi
     if [ -n "${RAC1PRIVIP1}" ] && [ -n "${RAC2PRIVIP1}" ] && [ -n "${RACPRIVFCNAME1}" ]; then
       cat <<EOF >>/etc/sysctl.conf
 net.ipv4.conf.${RACPRIVFCNAME1}.rp_filter = 2
@@ -2015,7 +2065,7 @@ EOF
   ##ROOT:
   if [ "$OS_VER_PRI" -eq 6 ]; then
     root_profile=/root/.profile
-  elif [ "$OS_VER_PRI" -eq 7 ]; then
+  elif [ "$OS_VER_PRI" -eq 7 ] || [ "$OS_VER_PRI" -eq 8 ]; then
     root_profile=/root/.bash_profile
   fi
   if [ "$(grep -E -c "#OracleBegin" ${root_profile})" -eq 0 ]; then
@@ -2036,7 +2086,7 @@ EOF
 EOF
     fi
   fi
-  logwrite "ROOT Profile" "cat ${root_profile}"
+  logwrite "ROOT Profile" "cat ${root_profile} | grep -v \"^\$\"|grep -v \"^#\""
 
   ##ORALCE:
   if [ "$(grep -E -c "#OracleBegin" /home/oracle/.bash_profile)" -eq 0 ]; then
@@ -2059,6 +2109,12 @@ alias sas='sqlplus / as sysdba'
 alias alert='tail -500f \$ORACLE_BASE/diag/rdbms/\$ORACLE_SID/\$ORACLE_SID/trace/alert_\$ORACLE_SID.log|more'
 export PS1="[\`whoami\`@\`hostname\`:"'\$PWD]\$ '
 EOF
+    ##Users are strongly recommended to go with 19.9 DB RU (or later) to minimize the number of Patches to be installed.19.9 OJVM & OCW RU Patches are also recommended to be applied,during/after the Installation.
+    if [ "${OS_VERSION}" = "linux8" ]; then
+      cat <<EOF >>/home/oracle/.bash_profile
+export CV_ASSUME_DISTID=OL7
+EOF
+    fi
     if rlwrap -v >/dev/null 2>&1; then
       cat <<EOF >>/home/oracle/.bash_profile
 alias sqlplus='rlwrap sqlplus'
@@ -2098,7 +2154,7 @@ EOF
       sed -i "s#ORACLE_HOME=${oracleHome}#ORACLE_HOME=${ENV_ORACLE_HOME}#" /home/oracle/.bash_profile
     fi
   fi
-  logwrite "Oracle Profile" "cat /home/oracle/.bash_profile"
+  logwrite "Oracle Profile" "cat /home/oracle/.bash_profile | grep -v \"^\$\"|grep -v \"^#\""
 
   ##GRID:
   if [ "${OracleInstallMode}" = "rac" ] || [ "${OracleInstallMode}" = "RAC" ] || [ "${OracleInstallMode}" = "restart" ] || [ "${OracleInstallMode}" = "RESTART" ]; then
@@ -2136,7 +2192,7 @@ EOF
 EOF
       fi
     fi
-    logwrite "Grid Profile" "cat /home/grid/.bash_profile"
+    logwrite "Grid Profile" "cat /home/grid/.bash_profile | grep -v \"^\$\"|grep -v \"^#\""
   fi
 }
 
@@ -2146,8 +2202,10 @@ EOF
 InstallRlwrap() {
   if [ "$(find "${SOFTWAREDIR}" -maxdepth 1 -name 'rlwrap-*.gz' | wc -l)" -gt 0 ]; then
     if ! rlwrap -v >/dev/null 2>&1; then
-      tar -zxvf "${SOFTWAREDIR}"/rlwrap* -C "${SOFTWAREDIR}"
-      cd "${SOFTWAREDIR}"/rlwrap-* || return
+      yum install -y tar
+      mkdir "${SOFTWAREDIR}"/rlwrap
+      tar -zxvf "${SOFTWAREDIR}"/rlwrap*tar.gz --strip-components 1 -C "${SOFTWAREDIR}"/rlwrap
+      cd "${SOFTWAREDIR}"/rlwrap || return
       ./configure && make && make install
     fi
     logwrite "rlwrap" "rlwrap -v"
@@ -2254,7 +2312,7 @@ UnzipGridSoft() {
         if [ "${OracleInstallMode}" = "rac" ] || [ "${OracleInstallMode}" = "RAC" ]; then
           scp "${ENV_GRID_HOME}"/cv/rpm/cvuqdisk-1.0.10-1.rpm "$RAC2HOSTNAME":"${SOFTWAREDIR}"
           ssh "$RAC2HOSTNAME" rpm -ivh "${SOFTWAREDIR}"/cvuqdisk-1.0.10-1.rpm
-          ssh "$RAC2HOSTNAME" rm -rf "${SOFTWAREDIR}"/cvuqdisk-1.0.9-1.rpm
+          ssh "$RAC2HOSTNAME" rm -rf "${SOFTWAREDIR}"/cvuqdisk-1.0.10-1.rpm
         fi
       else
         c1 "Make sure the cvuqdisk installation package is in the ${ENV_GRID_HOME}/cv/rpm directory:" red
@@ -2290,11 +2348,19 @@ Runcluvfy() {
     fi
   elif [[ "${DB_VERSION}" == "12.2.0.1" ]] || [[ "${DB_VERSION}" == "18.0.0.0" ]]; then
     if [ -f "${ENV_GRID_HOME}"/runcluvfy.sh ]; then
-      su - grid -c "${ENV_GRID_HOME}/runcluvfy.sh stage -pre crsinst -n $RAC1HOSTNAME,$RAC2HOSTNAME -fixup -verbose" | tee "${SOFTWAREDIR}"/runcluvfy.out
       # /tmp/$cvufix/runfixup.sh
       # if [ "${OracleInstallMode}" = "rac" ] || [ "${OracleInstallMode}" = "RAC" ]; then
       #     ssh "$RAC2HOSTNAME" /tmp/$cvufix/runfixup.sh
       # fi
+      EXPECT=/usr/bin/expect
+      USER_PROMPT="*# "
+      $EXPECT <<EOF
+      su - grid -c "${ENV_GRID_HOME}/runcluvfy.sh stage -pre crsinst -n $RAC1HOSTNAME,$RAC2HOSTNAME -fixup -verbose  -method root" | tee "${SOFTWAREDIR}"/runcluvfy.out
+      expect "*password*" {
+        send -- "$ROOTPASSWD\r"
+    } "*?assword:*" {send -- "$ROOTPASSWD\r"}
+expect "$USER_PROMPT"
+EOF
     fi
   elif [[ "${DB_VERSION}" == "19.3.0.0" ]]; then
     if [ -f "${ENV_GRID_HOME}"/runcluvfy.sh ]; then
@@ -3501,16 +3567,22 @@ EOF
 
   fi
 
+  if [ -d "/${SOFTWAREDIR}/"database ]; then
+    cd ~ || return
+    rm -rf "/${SOFTWAREDIR:?}/"database
+  fi
+
   if [ -f "${ENV_ORACLE_INVEN}"/orainstRoot.sh ] || [ -f "${ENV_ORACLE_HOME}"/root.sh ]; then
     if [ -f "${ENV_ORACLE_INVEN}"/orainstRoot.sh ]; then
       "${ENV_ORACLE_INVEN}"/orainstRoot.sh
+      if [ "${OracleInstallMode}" = "rac" ] || [ "${OracleInstallMode}" = "RAC" ]; then
+        ssh "$RAC2HOSTNAME" "${ENV_ORACLE_INVEN}"/orainstRoot.sh
+      fi
     fi
     if [ -f "${ENV_ORACLE_HOME}"/root.sh ]; then
       "${ENV_ORACLE_HOME}"/root.sh
       if [ "${OracleInstallMode}" = "rac" ] || [ "${OracleInstallMode}" = "RAC" ]; then
-        if [ "${OracleInstallMode}" = "rac" ] || [ "${OracleInstallMode}" = "RAC" ]; then
-          ssh "$RAC2HOSTNAME" "${ENV_ORACLE_HOME}"/root.sh
-        fi
+        ssh "$RAC2HOSTNAME" "${ENV_ORACLE_HOME}"/root.sh
       fi
     fi
   else
@@ -3606,16 +3678,16 @@ EOF
     if [ -d "/${SOFTWAREDIR}/""${GPATCH}" ] && [ -n "${GPATCH}" ]; then
       cd ~ || return
       rm -rf "/${SOFTWAREDIR:?}/""${GPATCH}"
+      ssh "$RAC2HOSTNAME" "/${SOFTWAREDIR:?}/""${GPATCH}"
     fi
   else
     if [ -d "/${SOFTWAREDIR}/""${OPATCH}" ] && [ -n "${OPATCH}" ]; then
       cd ~ || return
       rm -rf "/${SOFTWAREDIR:?}/""${OPATCH}"
+      if [ "${OracleInstallMode}" = "rac" ] || [ "${OracleInstallMode}" = "RAC" ]; then
+        ssh "$RAC2HOSTNAME" "/${SOFTWAREDIR:?}/""${OPATCH}"
+      fi
     fi
-  fi
-  if [ -d "/${SOFTWAREDIR}/"database ]; then
-    cd ~ || return
-    rm -rf "/${SOFTWAREDIR:?}/"database
   fi
 }
